updated enter calories section
