+++ requirements.txt
@@ -1,8 +1,7 @@
 bcrypt==3.2.0
 blinker==1.4
-bson==0.5.10
 certifi==1.0.1
-cffi==1.14.6
+cffi==1.15.1
 click==8.0.1
 colorama==0.4.4
 cryptography==3.4.8
@@ -19,7 +18,7 @@ Jinja2==3.0.1
 jwt==1.2.0
 MarkupSafe==2.0.1
 pycparser==2.20
-pymongo==3.11.3
+pymongo==4.5.0
 python-dateutil==2.8.2
 six==1.16.0
 tabulate==0.8.9
